prompt_generator/pyproject.toml
[project]
name = "prompt_generator"
version = "0.1.0"
description = "Interactive CLI tool for generating structured LLM prompts"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
